follow with axios done

+++ controllers/users.controller.js
@@ -19,6 +19,10 @@ module.exports.profile = (req, res, next) => {
             Post.find({ creator: userId })
             .populate('creator')
             .then((posts) => {
+                posts.forEach((post) => {
+                    const postCreator = post.creator.id.valueOf();
+                    return post.sameOwner = currentUser.id === postCreator ? true : false;
+                })
                 posts.reverse()
                 res.render('users/profile', { user, posts, imFollower, itsMeMario })
             })
@@ -44,25 +48,28 @@ module.exports.doFollow = (req, res, next) => {
     const currentUser = req.user;
 
     User.findOne({username}, {id: 1})
-    .then((userID) => {        
-        Follow.create({follower: currentUser.id, following: userID.id})
-        .then((followCreated) => {
-            res.redirect(`/profile/${username}`);
-        })
-    })    
-    .catch((err) => next(err))
-};
-
-module.exports.doUnfollow = (req, res, next) => {
-    const { username } = req.params;
-    const currentUser = req.user;
-
-    User.findOne({username}, {id: 1})
-    .then((userID) => {        
-        Follow.findOneAndDelete({follower: currentUser.id, following: userID.id})
-        .then((followDeleted) => {
-            res.redirect(`/profile/${username}`);
-        })
+    .then((userID) => {  
+        console.log('userID:', currentUser.id)
+        Follow.findOne({follower: currentUser.id, following: userID.id})
+        .then((response) => {
+            console.log('response:', response)
+            if(response) {
+                Follow.findOneAndDelete({follower: currentUser.id, following: userID.id})
+                .then((followDeleted) => {
+                    console.log(followDeleted)
+                    res.status(204).send(followDeleted);
+                })
+                .catch((err) => next(err))
+            } else {
+                Follow.create({follower: currentUser.id, following: userID.id})
+                .then((followCreated) => {
+                    console.log(followCreated)
+                    res.status(204).send(followCreated);
+                })
+                .catch((err) => next(err))
+            }
+        })      
+        
     })    
     .catch((err) => next(err))
 };
